refactor(web): split PlaylistBuilderPanel and ShowsPanel into partner dirs
Same shape as the LibraryPanel split, applied across the sibling set so
the admin directory doesn't end up half-converted.

PlaylistBuilderPanel 1739 -> 1388, with components/admin/playlist-builder/:
  types.ts          shapes, vocab, the range bounds and formatters
  generate.ts       the poll-a-server-side-job generation driver
  bits.tsx          Eyeb/Tog/IconBtn/Chip/DualRange/SwitchRow + energy scale
  EnergyGraph.tsx   the energy tape-strip

ShowsPanel 1696 -> 619, with components/admin/shows/:
  types.ts          shapes and the caps that mirror settings.ts
  lib.ts            hydrate / validate / project helpers (mirrors schedule/lib.ts)
  ChipRow.tsx       the multi-value filter chips
  ShowAvatar.tsx    the initials-behind-<img> avatar
  ShowEditor.tsx    the inline editor (550 lines on its own)
  ShowDefRow.tsx    one row of the definitions list

ShowsTable.tsx and ShowPickers.tsx moved into shows/ too. Both were
imported only by ShowsPanel, so leaving them at the admin root while
their siblings moved would have been the worst of both layouts.

hydrateShow is worth calling out: it's the single place the legacy
singular -> plural music-filter coercion (#929) happens, and it now sits
in lib.ts next to showValid/showPayload rather than buried between two
components, so the initial load and a community install visibly share it.

No behaviour change. Web lint clean and `next build` succeeds.
