+++ web/components/admin/PlaylistBuilderPanel.tsx
@@ -21,395 +21,44 @@ import { Switch } from '../ui/switch';
 import { V3Alert } from '../ui/alert';
 import { ScrollArea } from '../ui/scroll-area';
 import { cn } from '../../lib/cn';
-
-const API = (process.env.NEXT_PUBLIC_API_URL as string | undefined) || '/api';
-
-// Mirrors SHOW_MOODS in controller/src/settings.ts (stable vocab).
-const MOODS = [
-  'energetic', 'calm', 'reflective', 'celebratory', 'romantic', 'spiritual',
-  'focus', 'workout', 'driving', 'cooking', 'rainy', 'sunny', 'night', 'morning',
-  'evening', 'festival', 'cultural',
-];
-const ENERGIES = ['low', 'medium', 'high'];
-type ArcShape = 'flat' | 'build' | 'peak-then-cool' | 'wind-down';
-const ARCS: { id: ArcShape; label: string; hint: string }[] = [
-  { id: 'flat', label: 'Steady', hint: 'even energy throughout' },
-  { id: 'build', label: 'Build', hint: 'calm → energetic' },
-  { id: 'peak-then-cool', label: 'Peak', hint: 'rise, then cool down' },
-  { id: 'wind-down', label: 'Wind down', hint: 'high → mellow' },
-];
-// Band domains. Anchors parked at the extremes mean "unbounded" on that end.
-const LEN_MAX = 600;                              // track length: 0 → 10:00, 15s notches
-const LEN_STEP = 15;
-const BPM_MIN = 60;                               // tempo: 60 → 200 bpm, 5 bpm notches
-const BPM_MAX = 200;
-const BPM_STEP = 5;
-const YEAR_MIN = 1950;                            // release year: 1950 → current year
-const YEAR_MAX = new Date().getFullYear();
-
-// Bar palette for the energy graph — theme-aware mixes rather than the mock's
-// light-theme hexes, so dark mode keeps the same low/med/high contrast. Raw
-// values feed SVG `fill` attributes; the class twins style HTML swatches.
-const EN_LOW = 'color-mix(in oklab, var(--ink) 22%, var(--bg))';
-const EN_MED = 'color-mix(in oklab, var(--ink) 80%, var(--bg))';
-const EN_HIGH = 'var(--accent)';
-const EN_LOW_BG = 'bg-[color-mix(in_oklab,var(--ink)_22%,var(--bg))]';
-const EN_MED_BG = 'bg-[color-mix(in_oklab,var(--ink)_80%,var(--bg))]';
-const EN_HIGH_BG = 'bg-[var(--accent)]';
-
-type View = 'result' | 'empty' | 'generating' | 'nomatch' | 'error';
-type GenMode = 'fresh' | 'regenerate' | 'more';
-
-interface DraftTrack {
-  id: string;
-  title: string;
-  artist: string;
-  album?: string;
-  durationSec: number;
-  year?: number | null;
-  genre?: string | null;
-  energy?: string | null;
-  moods?: string[];
-  instrumental?: boolean | null;
-}
-interface SeedChip { id: string; title: string; artist: string }
-interface PlaylistSummary { id: string; name: string; songCount: number; synced?: boolean; lastSyncedAt?: string | null }
-
-// Loose shape for /dj/search rows and /playlists/:id entries — the controller
-// returns Subsonic-derived fields with varying key names across endpoints.
-interface RawTrackRow {
-  id: string;
-  title?: string;
-  artist?: string;
-  album?: string;
-  duration?: number;
-  durationSec?: number;
-  year?: number | null;
-  genre?: string | null;
-  energy?: string | null;
-  moods?: string[];
-}
-
-function rowToDraft(s: RawTrackRow): DraftTrack {
-  return {
-    id: s.id,
-    title: s.title || '',
-    artist: s.artist || '',
-    album: s.album,
-    durationSec: s.durationSec ?? s.duration ?? 0,
-    year: s.year,
-    genre: s.genre ?? null,
-    energy: s.energy ?? null,
-    moods: s.moods || [],
-    instrumental: null,
-  };
-}
-
-function fmtDur(sec: number): string {
-  const s = Math.max(0, Math.round(sec));
-  const h = Math.floor(s / 3600);
-  const m = Math.floor((s % 3600) / 60);
-  const ss = s % 60;
-  if (h) return `${h}:${String(m).padStart(2, '0')}:${String(ss).padStart(2, '0')}`;
-  return `${m}:${String(ss).padStart(2, '0')}`;
-}
-function fmtRun(total: number): string {
-  const h = Math.floor(total / 3600);
-  const m = Math.round((total % 3600) / 60);
-  return h > 0 ? `${h}h ${m}m` : `${m}m`;
-}
-function relTime(iso: string): string {
-  const mins = Math.floor((Date.now() - new Date(iso).getTime()) / 60000);
-  if (mins < 1) return 'just now';
-  if (mins < 60) return `${mins}m ago`;
-  const h = Math.floor(mins / 60);
-  if (h < 24) return `${h}h ago`;
-  const d = Math.floor(h / 24);
-  if (d < 30) return `${d}d ago`;
-  return new Date(iso).toLocaleDateString();
-}
-// ── generation over a server-side job ────────────────────────────────────────
-// A generation legitimately runs for minutes, and Cloudflare cuts proxied
-// responses off at ~100s with an HTML error page — so the panel starts a job
-// (POST /playlists/generate/jobs) and polls it instead of holding one request
-// open. Bodies are also never parsed before checking they ARE JSON: WebKit
-// reports r.json() on that HTML page as the baffling "The string did not
-// match the expected pattern".
-
-type AdminFetch = (path: string, init?: RequestInit) => Promise<Response>;
-
-async function readJsonSafe(r: Response): Promise<any> {
-  const ct = r.headers.get('content-type') || '';
-  if (!ct.includes('application/json')) {
-    throw new Error(`the curation service returned an unexpected response (HTTP ${r.status}) — is the controller reachable?`);
-  }
-  try {
-    return await r.json();
-  } catch {
-    throw new Error(`the curation service returned malformed JSON (HTTP ${r.status})`);
-  }
-}
-
-const sleep = (ms: number) => new Promise<void>((resolve) => setTimeout(resolve, ms));
-
-const GEN_POLL_MS = 2000;
-const GEN_DEADLINE_MS = 10 * 60_000;
-const GEN_POLL_MISSES = 3; // consecutive transient poll failures tolerated
-
-// Resolves with the GenerateResult payload, throws with an operator-readable
-// message otherwise.
-async function runGenerationJob(adminFetch: AdminFetch, body: unknown): Promise<any> {
-  const init: RequestInit = {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(body),
-  };
-  const start = await adminFetch('/playlists/generate/jobs', init);
-  // A pre-jobs controller 404s here (mid-upgrade version skew) — fall back to
-  // the synchronous endpoint rather than failing the click.
-  if (start.status === 404) {
-    const r = await adminFetch('/playlists/generate', init);
-    const j = await readJsonSafe(r);
-    if (!r.ok) throw new Error(j.error || 'generation failed');
-    return j;
-  }
-  const started = await readJsonSafe(start);
-  if (!start.ok || !started.jobId) throw new Error(started.error || 'generation failed to start');
-  const deadline = Date.now() + GEN_DEADLINE_MS;
-  let misses = 0;
-  while (Date.now() < deadline) {
-    await sleep(GEN_POLL_MS);
-    let poll: any;
-    try {
-      const r = await adminFetch(`/playlists/generate/jobs/${started.jobId}`);
-      poll = await readJsonSafe(r);
-      if (!r.ok) throw new Error(poll.error || `poll failed (HTTP ${r.status})`);
-    } catch (err) {
-      if (++misses >= GEN_POLL_MISSES) throw err instanceof Error ? err : new Error('lost contact with the curation service');
-      continue;
-    }
-    misses = 0;
-    if (poll.status === 'running') continue;
-    if (poll.status === 'error') throw new Error(poll.error || 'generation failed');
-    return poll.result || {};
-  }
-  throw new Error('generation is taking unusually long — it may still land server-side; try again in a minute');
-}
-
-const energyPct = (e?: string | null): number => (e === 'low' ? 34 : e === 'high' ? 92 : 64);
-const energyColor = (e?: string | null): string => (e === 'low' ? EN_LOW : e === 'high' ? EN_HIGH : EN_MED);
-const energyBgClass = (e?: string | null): string => (e === 'low' ? EN_LOW_BG : e === 'high' ? EN_HIGH_BG : EN_MED_BG);
-// Untagged tracks read '—', not a fake 'med' — an untagged library shouldn't
-// masquerade as uniformly mid-energy (the bars go translucent for the same reason).
-const energyLabel = (e?: string | null): string =>
-  e === 'low' || e === 'medium' || e === 'high' ? (e === 'medium' ? 'med' : e) : '—';
-const energyKnown = (e?: string | null): boolean => e === 'low' || e === 'medium' || e === 'high';
-
-// ── shared micro-pieces (design idiom: mono eyebrows, sharp toggles) ─────────
-
-function Eyeb({ children, muted, className }: { children: React.ReactNode; muted?: boolean; className?: string }) {
-  return (
-    <span className={cn('font-mono text-[10px] font-bold tracking-[0.16em] uppercase', muted ? 'text-muted' : 'text-ink', className)}>
-      {children}
-    </span>
-  );
-}
-
-function Tog({ on, onClick, title, children }: { on: boolean; onClick: () => void; title?: string; children: React.ReactNode }) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      title={title}
-      className={cn(
-        'border px-[11px] py-1.5 font-mono text-[11px] font-semibold tracking-[0.03em] transition',
-        on ? 'border-ink bg-ink text-bg' : 'border-separator-strong bg-bg text-ink hover:border-ink',
-      )}
-    >
-      {children}
-    </button>
-  );
-}
-
-function IconBtn({ onClick, disabled, title, className, children }: {
-  onClick?: () => void; disabled?: boolean; title?: string; className?: string; children: React.ReactNode;
-}) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      disabled={disabled}
-      title={title}
-      className={cn(
-        'grid size-[30px] place-items-center border border-transparent text-muted transition',
-        'hover:border-separator-soft hover:bg-ink-soft hover:text-ink disabled:opacity-25 disabled:hover:border-transparent disabled:hover:bg-transparent',
-        className,
-      )}
-    >
-      {children}
-    </button>
-  );
-}
-
-function Chip({ accent, onRemove, children }: { accent?: boolean; onRemove?: () => void; children: React.ReactNode }) {
-  return (
-    <span className={cn(
-      'inline-flex items-center gap-1.5 border bg-bg px-[7px] py-[3px] font-mono text-[10px] font-semibold tracking-[0.06em] uppercase',
-      accent ? 'border-[var(--accent)] text-vermilion' : 'border-separator-strong text-ink',
-    )}>
-      {children}
-      {onRemove && (
-        <button type="button" onClick={onRemove} className="cursor-pointer text-muted hover:text-ink" title="remove">
-          <X className="size-3" />
-        </button>
-      )}
-    </span>
-  );
-}
-
-// ── Dual-anchor range — two overlaid native sliders sharing one track, with an
-// accent band between the anchors. No dependency; thumbs stay keyboardable.
-function DualRange({ min, max, step, lo, hi, disabled, onLo, onHi, loLabel, hiLabel }: {
-  min: number; max: number; step: number; lo: number; hi: number; disabled?: boolean;
-  onLo: (v: number) => void; onHi: (v: number) => void; loLabel: string; hiLabel: string;
-}) {
-  const bandRef = useRef<HTMLDivElement>(null);
-  const span = max - min || 1;
-  const loPct = ((lo - min) / span) * 100;
-  const hiPct = ((hi - min) / span) * 100;
-  useDynamicStyle(bandRef, { left: `${loPct}%`, width: `${Math.max(0, hiPct - loPct)}%` });
-  const thumb =
-    'pointer-events-none absolute inset-0 h-5 w-full appearance-none bg-transparent outline-none ' +
-    '[&::-webkit-slider-runnable-track]:bg-transparent [&::-webkit-slider-thumb]:pointer-events-auto ' +
-    '[&::-webkit-slider-thumb]:size-3.5 [&::-webkit-slider-thumb]:appearance-none ' +
-    '[&::-webkit-slider-thumb]:border [&::-webkit-slider-thumb]:border-ink [&::-webkit-slider-thumb]:bg-[var(--accent)] ' +
-    '[&::-moz-range-track]:bg-transparent [&::-moz-range-thumb]:pointer-events-auto [&::-moz-range-thumb]:size-3.5 ' +
-    '[&::-moz-range-thumb]:rounded-none [&::-moz-range-thumb]:border [&::-moz-range-thumb]:border-ink [&::-moz-range-thumb]:bg-[var(--accent)]';
-  return (
-    <div className={cn('relative h-5', disabled && 'opacity-40')}>
-      <div className="absolute top-1/2 right-0 left-0 h-[3px] -translate-y-1/2 bg-separator-strong" />
-      <div ref={bandRef} className="absolute top-1/2 h-[3px] -translate-y-1/2 bg-[var(--accent)]" />
-      <input
-        type="range" min={min} max={max} step={step} value={lo} disabled={disabled}
-        onChange={e => onLo(Math.min(+e.target.value, hi))}
-        aria-label={loLabel}
-        // When both anchors crowd the right end, lift the lo thumb so it stays grabbable.
-        className={cn(thumb, lo > max - step * 4 && 'z-10')}
-      />
-      <input
-        type="range" min={min} max={max} step={step} value={hi} disabled={disabled}
-        onChange={e => onHi(Math.max(+e.target.value, lo))}
-        aria-label={hiLabel}
-        className={thumb}
-      />
-    </div>
-  );
-}
-
-function SwitchRow({ label, hint, on, onToggle, mutedLabel }: {
-  label: string; hint: string; on: boolean; onToggle: (v: boolean) => void; mutedLabel?: boolean;
-}) {
-  return (
-    <div className="flex items-center justify-between gap-3">
-      <div>
-        <div className={cn('text-[13px] font-semibold', mutedLabel && 'text-muted')}>{label}</div>
-        <div className="font-mono text-[10px] text-muted">{hint}</div>
-      </div>
-      <Switch checked={on} onCheckedChange={onToggle} aria-label={label} />
-    </div>
-  );
-}
-
-// ── Energy tape-strip — slim per-track bars + dashed target arc. Collapsible,
-// and every bar is a jump-link: click scrolls its track row into view. ────────
-
-function EnergyGraph({ tracks, arc, open, onToggle, onBarClick }: {
-  tracks: DraftTrack[]; arc: ArcShape; open: boolean; onToggle: () => void; onBarClick: (i: number) => void;
-}) {
-  const n = tracks.length;
-  const arcLabel = ARCS.find(a => a.id === arc)?.label || 'Steady';
-  const noneTagged = useMemo(() => tracks.every(t => !energyKnown(t.energy)), [tracks]);
-  const targetPts = useMemo(() => {
-    const f = (p: number): number => {
-      if (arc === 'build') return p;
-      if (arc === 'wind-down') return 1 - p;
-      if (arc === 'peak-then-cool') return p < 0.6 ? p / 0.6 : 1 - ((p - 0.6) / 0.4) * 0.65;
-      return 0.5;
-    };
-    return tracks.map((_, i) => {
-      const p = n > 1 ? i / (n - 1) : 0;
-      return `${(i + 0.5).toFixed(2)},${(82 - f(p) * 64).toFixed(2)}`;
-    }).join(' ');
-  }, [tracks, arc, n]);
-  if (n === 0) return null;
-  return (
-    <div className="flex-none border-b border-separator-soft px-4 sm:px-6">
-      <button
-        type="button"
-        onClick={onToggle}
-        className="flex h-7 w-full items-center justify-between gap-3 text-left"
-        title={open ? 'collapse the energy strip' : 'expand the energy strip'}
-      >
-        <span className="flex min-w-0 items-center gap-2.5">
-          <Eyeb muted>Energy</Eyeb>
-          {noneTagged && open && (
-            <span className="truncate font-mono text-[9px] text-muted/80">
-              untagged — run the library tagger (Library → Tag) to chart the real arc
-            </span>
-          )}
-        </span>
-        <span className="flex flex-none items-center gap-3 font-mono text-[9px] text-muted">
-          {open && !noneTagged && (
-            <>
-              <span className="hidden items-center gap-1 sm:flex"><span className={cn('inline-block size-2', EN_LOW_BG)} />low</span>
-              <span className="hidden items-center gap-1 sm:flex"><span className={cn('inline-block size-2', EN_MED_BG)} />med</span>
-              <span className="hidden items-center gap-1 sm:flex"><span className={cn('inline-block size-2', EN_HIGH_BG)} />high</span>
-            </>
-          )}
-          {open && (
-            <span className="flex items-center gap-1">
-              <svg width="14" height="8" aria-hidden><line x1="0" y1="4" x2="14" y2="4" stroke="var(--ink)" strokeWidth="1.5" strokeDasharray="3 2" /></svg>
-              target · {arcLabel}
-            </span>
-          )}
-          {open ? <ChevronUp className="size-3.5" /> : <ChevronDown className="size-3.5" />}
-        </span>
-      </button>
-      {open && (
-        <div className="relative h-11 pb-1.5">
-          <svg viewBox={`0 0 ${n} 100`} preserveAspectRatio="none" className="block h-full w-full">
-            {tracks.map((t, i) => {
-              const pct = energyPct(t.energy);
-              return (
-                <rect
-                  key={`${t.id}-${i}`}
-                  x={(i + 0.1).toFixed(3)}
-                  y={(100 - pct).toFixed(2)}
-                  width={0.8}
-                  height={pct}
-                  fill={energyColor(t.energy)}
-                  onClick={() => onBarClick(i)}
-                  className={cn('cursor-pointer hover:opacity-70', !energyKnown(t.energy) && 'opacity-35')}
-                >
-                  <title>{i + 1}. {t.title} — {t.artist}</title>
-                </rect>
-              );
-            })}
-            <polyline
-              points={targetPts}
-              fill="none"
-              stroke="var(--ink)"
-              strokeWidth="1.5"
-              strokeDasharray="3 2"
-              vectorEffect="non-scaling-stroke"
-              className="pointer-events-none opacity-60"
-            />
-          </svg>
-        </div>
-      )}
-    </div>
-  );
-}
+import { EnergyGraph } from './playlist-builder/EnergyGraph';
+import {
+  Chip,
+  DualRange,
+  Eyeb,
+  IconBtn,
+  SwitchRow,
+  Tog,
+  energyBgClass,
+  energyLabel,
+} from './playlist-builder/bits';
+import { runGenerationJob } from './playlist-builder/generate';
+import type {
+  ArcShape,
+  DraftTrack,
+  GenMode,
+  PlaylistSummary,
+  RawTrackRow,
+  SeedChip,
+  View,
+} from './playlist-builder/types';
+import {
+  API,
+  ARCS,
+  BPM_MAX,
+  BPM_MIN,
+  BPM_STEP,
+  ENERGIES,
+  LEN_MAX,
+  LEN_STEP,
+  MOODS,
+  YEAR_MAX,
+  YEAR_MIN,
+  fmtDur,
+  fmtRun,
+  relTime,
+  rowToDraft,
+} from './playlist-builder/types';
 
 // ─────────────────────────────────────────────────────────────────────────────
